Add files via upload
This was an interesting one, as my initial attempts took far too long to run. At first, I figured my algorithm was bad, so I tried a couple of different things. However, what I found was that it was my casual use of the python .remove() method that was causing the program to search through the list of 2 million items each time (granted that size gradually reduced). By changing the program to use the same algorithm but without the .remove() method, the program produced the solution in seconds.